Add collapsed json file, Add tests for JsonTranslatorTest, Add methods for JsonTranslator
There are now different example files for collapsed JSON (useful for pasting) and Pretty Printed JSON (useful for finding individual values).
Since we will need users and timestamps, tests were added to fetch them as a List and as a String.
As such, JsonTranslator has two new methods, is no longer static, and stores the JSON String used to construct it as an `Object document`. This is because, when reading from a string, JsonPath has to parse the JSON every time. Storing the parsed JSON in a document evades this issue.