Refactor JSON files, remove redundant tests that call List<String>.toString(), fix platform incompatibility with WebPageDataTest, add test and method for fetching a single string from a list from a JSON list.
Renamed JSON files and removed compressed JSON file. We are storing the resources with prettified print, and compressing it only when pasting it into code.
There is no reason to return List<String>.toString. Return the list, and the object asking for said list can use the .toString() from there.
On Cameron's MacBook, the WebPageDataTest worked just fine. On Windows Laptop, it would inexplicably fail. After some debugging, I discovered that the difference on my end was the use of '\n' and '\r\n' (aka '%n' in String.format('%n'). Changing the expected list's line breaks from \n to %n rectified the issue.
We discovered that, when using a JsonPath operator other than a deep scan, when it returned a JSON block {"this": "is", "that": "at"}, it is stored as a LinkedHashMap<String>. A call to <JSON Block>.this returns the value, and "this" is the key.

This is too much work for a single commit, and we will endeavor to avoid this mistake in the future.