Rewrite error handling
Instead of WikiApiFetcher throwing errors on behalf of the WebPageData, the WebPageData can carry a "problem" that is then checked right before said WebPageData is printed in Main.

You may notice that this means a lot of methods no longer throw extraneous exceptions.

One last note: If one worries about this commit's timing, we did all of this after class, but I forgot to commit it.